add ctrl+k as shortcut for search chat list
