+++ src/renderer/App.js
@@ -6,39 +6,14 @@ import { addLocaleData, IntlProvider } from 'react-intl'
 import enLocaleData from 'react-intl/locale-data/en'
 import { remote } from 'electron'
 import { callDcMethod, sendToBackend, sendToBackendSync, ipcBackend, startBackendLogging } from './ipc'
+import attachKeybindingsListener from './keybindings'
 
 const localize = require('../localize')
 const moment = require('moment')
 
 addLocaleData(enLocaleData)
 
-function selectFirstChatListItem() {
-  let chatItemToSelect = document.querySelector('.chat-list-item')
-  if (chatItemToSelect.classList.contains('chat-list-item--is-deaddrop')) {
-    chatItemToSelect = chatItemToSelect.nextSibling
-  }
-  chatItemToSelect.click()
-}
-
-function selectChatItem(domChatItem) {
-  if (domChatItem.classList.contains('chat-list-item--is-deaddrop')) return
-  domChatItem.click()
-  domChatItem.scrollIntoView({block: 'nearest'})
-}
-
-document.addEventListener ("keydown", function (Event) {
-  if (Event.altKey  &&  Event.key === "ArrowDown") {
-    const selectedChatItems = document.getElementsByClassName('chat-list-item--is-selected')
-    if (selectedChatItems.length == 0) return selectFirstChatListItem()
-    const nextChatItem = selectedChatItems[0].nextSibling
-    selectChatItem(nextChatItem)
-  } else if(Event.altKey && Event.key === 'ArrowUp') {
-    const selectedChatItems = document.getElementsByClassName('chat-list-item--is-selected')
-    if (selectedChatItems.length == 0) return selectFirstChatListItem()
-    const previousChatItem = selectedChatItems[0].previousSibling
-    selectChatItem(previousChatItem)
-  }
-});
+attachKeybindingsListener()
 
 export default function App (props) {
   const [state, setState] = useState(remote.app.state)
